+++ packages/chain-consumer/src/index.ts
@@ -1,4 +1,5 @@
 export * from './consumers'
 export * from './decomposers'
 export * from './composers'
+export * from './proto-composers'
 export * from './types'