+++ packages/chain-consumer/src/proto-composers/index.ts
@@ -1,0 +1,6 @@
+export * from './DistributionProtoComposer'
+export * from './GovernanceProtoComposer'
+export * from './PeggyProtoComposer'
+export * from './StakingProtoComposer'
+export * from './InsuranceProtoComposer'
+export * from './BankProtoComposer'